requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.9